Update content-page-full-width.blade.php
I propose to add the attribute "u-shift--left--1-col--at-xxlarge" in the section id="top" div.  This attribute is missing in the original wordpress template on https://www.adventist.io/themes/wordpress/  leading to the initial setup having a shifted section aligning on the left edge of the screen. Each time I setup a new website I always add that attribute to solve the problem, so best to make changes to the original download.